split direct build gated dependency tests
