+++ docs/COMPLETION_AUDIT.md
@@ -311,6 +311,10 @@ and do not assume Phase 4 is ready without evidence.
   test dependency rejection coverage into
   `tests/integration/cli_build/build_command_validation/gated_dependencies.rs`,
   leaving executable-target and skipped-source checks in the parent module.
+- Direct `zen build.zen` validation tests now split direct and transitive gated
+  test dependency rejection coverage into
+  `tests/integration/cli_build/direct_build_graph_validation/gated_dependencies.rs`,
+  matching the normal build-command validation layout.
 - Resolver-backed behavior impl metadata now builds restored impl-block tasks
   once and reuses them for both impl method signature restoration and omitted
   default-method synthesis, preserving the signature-before-defaults ordering.
