+++ docs/PHASE_PLAN.md
@@ -487,6 +487,9 @@ checked-in docs, tests, and commits only.
   test dependency rejection coverage in a focused child module, leaving the
   parent build command validation file for executable-target and skipped-source
   checks.
+- Direct `zen build.zen` validation tests now keep direct and transitive gated
+  test dependency rejection coverage in a focused child module, matching the
+  normal build-command validation layout.
 - Resolver-backed behavior impl metadata now builds restored impl-block tasks
   once and reuses them for both impl method signature restoration and omitted
   default-method synthesis. Impl-block declaration collection now also uses a
